requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.1.0
exceptiongroup==1.1.0
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.21.0
outcome==1.2.0
pandas==1.5.2
pip==22.3.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pymongo==3.11.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
scikit-learn==1.1.2
scipy==1.3.2
selenium==4.7.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==2.0.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wsproto==1.2.0
xgboost==1.7.2
zipp==3.11.0